feat: Add data storage business logic

+++ backend/src/main/java/com/example/demoio/modules/datastorage/dto/SaveGameData.java
@@ -1,4 +1,4 @@
 package com.example.demoio.modules.datastorage.dto;
 
-public record SaveGameData(int gameId, String gameDataJSON) {
+public record SaveGameData(Long gameId, String gameDataJSON) {
 }
